+++ packages/components/src/ProTable/BaseTable.tsx
@@ -37,7 +37,6 @@ const BaseTable: React.FC<BaseTableProps> = ({
     searchValues,
     SWRConfiguration = {},
   } = store as any;
-
   const { selectKey, type = 'checkbox', defaultSelected = [] } = rowSelection;
   const { x } = scroll;
 
@@ -77,7 +76,6 @@ const BaseTable: React.FC<BaseTableProps> = ({
         defaultSearchValues[name] = col.props.initialValue;
       }
     });
-    (store as any).updateSearchValueRef.current = defaultSearchValues || {};
     return defaultSearchValues;
   }, [JSON.stringify(columns)]);
 
@@ -100,15 +98,19 @@ const BaseTable: React.FC<BaseTableProps> = ({
   }, [pageIndex, JSON.stringify(defaultValues), JSON.stringify(searchValues)]);
 
   const pageSize = formatQuery().pageSize || 10;
-
-  const { data, isValidating } = useSWR(
+  const { data, isValidating, mutate } = useSWR(
     [key, { method: 'POST', body: formatQuery() }],
     request,
     {
       revalidateOnFocus: false,
+      revalidateOnMount: false,
       ...SWRConfiguration,
     },
   );
+
+  useEffect(() => {
+    mutate(false);
+  }, [mutate]);
 
   // table数据
   const tableData =
@@ -125,24 +127,18 @@ const BaseTable: React.FC<BaseTableProps> = ({
   );
 
   // 查询
-  const onSearch = () => {
-    setPageIndex(1);
-    // 更新searchValues,触发请求接口
-    updateStore({
-      searchValues: {
-        ...(store as any).updateSearchValueRef.current,
-        swr_Rest_Time: new Date().getTime(),
-      },
-    });
+  const onSearch = async () => {
+    await setPageIndex(1);
+    mutate(false);
   };
-
   // 分页
   const onPageChange = useCallback(
-    (page) => {
+    async (page) => {
       if (pageChange) {
         pageChange(page);
       }
-      setPageIndex(page);
+      await setPageIndex(page);
+      mutate(false);
     },
     [setPageIndex, pageChange],
   );
@@ -159,7 +155,7 @@ const BaseTable: React.FC<BaseTableProps> = ({
     const stores: any = {
       data: data?.data,
       total: data?.total,
-      loading: !data || isValidating,
+      loading: isValidating,
       onSearch,
       selection,
       pageIndex,
@@ -168,7 +164,6 @@ const BaseTable: React.FC<BaseTableProps> = ({
     if (!isFirstMountRef.current) {
       isFirstMountRef.current = true;
       // 默认表单值
-
       stores.searchValues = defaultSearchValues;
     }