+++ CHANGELOG.md
@@ -5,11 +5,19 @@
 严格遵循 [`Semantic Versioning 2.0.0`](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
 
-## 5.3.0
+## 5.2.10
 
 - @uiw-admin/components
   - 🐞`ProForm`:增加`cardProps` `collapseProps` `collapsePanelProps` api 可配置card collapse参数
 
+- @uiw-admin/components
+  - 🐞`ProDrawer`:修复文档报错
+  - 🐞`ProTable`: 增加 `scroll` api
+  - 🐞`ProTable`: FormCol 增加 `align`设置列的对齐方式 api
+  - 🐞`ProTable`: 修复loading一直存在问题
+- @uiw-admin/utils
+  - 🐞 修复`request`请求参数问题
+  
 ## 5.2.9
 
 - @uiw-admin/components